requirements.txt
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.7.14
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cache2==0.2.2
fastapi-versioning==0.10.0
flower==2.0.1
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
redis==6.2.0
rsa==4.9.1
ruff==0.12.4
sentry-sdk==2.33.0
six==1.17.0
sniffio==1.3.1
sqladmin==0.21.0
SQLAlchemy==2.0.41
starlette==0.47.1
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
WTForms==3.1.2
